requirements.txt
Flask==3.0.3
youtube-transcript-api==0.6.2
google-generativeai==0.8.3
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.4
requests==2.31.0
urllib3==2.0.7
